fix(address): demo拆解与规范
